finish moving instances over to wasmtime_wasi_io
