+++ crates/wasi/src/lib.rs
@@ -304,15 +304,16 @@ pub fn add_to_linker_with_options_async<T: WasiView>(
 ) -> anyhow::Result<()> {
     let l = linker;
     let io_closure = io_type_annotate::<T, _>(|t| IoImpl(t));
+    wasmtime_wasi_io::bindings::wasi::io::error::add_to_linker_get_host(l, io_closure)?;
+    wasmtime_wasi_io::bindings::wasi::io::poll::add_to_linker_get_host(l, io_closure)?;
+    wasmtime_wasi_io::bindings::wasi::io::streams::add_to_linker_get_host(l, io_closure)?;
+
     let closure = type_annotate::<T, _>(|t| WasiImpl(IoImpl(t)));
 
     crate::bindings::clocks::wall_clock::add_to_linker_get_host(l, closure)?;
     crate::bindings::clocks::monotonic_clock::add_to_linker_get_host(l, closure)?;
     crate::bindings::filesystem::types::add_to_linker_get_host(l, closure)?;
     crate::bindings::filesystem::preopens::add_to_linker_get_host(l, closure)?;
-    crate::bindings::io::error::add_to_linker_get_host(l, io_closure)?;
-    crate::bindings::io::poll::add_to_linker_get_host(l, io_closure)?;
-    crate::bindings::io::streams::add_to_linker_get_host(l, io_closure)?;
     crate::bindings::random::random::add_to_linker_get_host(l, closure)?;
     crate::bindings::random::insecure::add_to_linker_get_host(l, closure)?;
     crate::bindings::random::insecure_seed::add_to_linker_get_host(l, closure)?;
@@ -405,15 +406,17 @@ pub fn add_to_linker_with_options_sync<T: WasiView>(
 ) -> anyhow::Result<()> {
     let l = linker;
     let io_closure = io_type_annotate::<T, _>(|t| IoImpl(t));
+    wasmtime_wasi_io::bindings::wasi::io::error::add_to_linker_get_host(l, io_closure)?;
+
+    crate::bindings::sync::io::poll::add_to_linker_get_host(l, io_closure)?;
+    crate::bindings::sync::io::streams::add_to_linker_get_host(l, io_closure)?;
+
     let closure = type_annotate::<T, _>(|t| WasiImpl(IoImpl(t)));
 
     crate::bindings::clocks::wall_clock::add_to_linker_get_host(l, closure)?;
     crate::bindings::clocks::monotonic_clock::add_to_linker_get_host(l, closure)?;
     crate::bindings::sync::filesystem::types::add_to_linker_get_host(l, closure)?;
     crate::bindings::filesystem::preopens::add_to_linker_get_host(l, closure)?;
-    crate::bindings::io::error::add_to_linker_get_host(l, io_closure)?;
-    crate::bindings::sync::io::poll::add_to_linker_get_host(l, io_closure)?;
-    crate::bindings::sync::io::streams::add_to_linker_get_host(l, io_closure)?;
     crate::bindings::random::random::add_to_linker_get_host(l, closure)?;
     crate::bindings::random::insecure::add_to_linker_get_host(l, closure)?;
     crate::bindings::random::insecure_seed::add_to_linker_get_host(l, closure)?;
